+++ changelog.md
@@ -2,6 +2,7 @@
 ### New / Changed Widgets
 
 ### Other New Features
+- calendar.list has an option to show event links and locations in collapsible areas scrolling down on click.  
 
 ### Improvements
 - performance increased by avoiding long executions times of frequent jQuery mobile pagecontainer widget calls
@@ -17,6 +18,7 @@
 - calendar names in config did not allow whitespace after comma
 - clock.digiclock always had double events
 - widget assistant deleted whitespaces even within the parameters (e.g. calendar names containing whistespaces wont't work)
+- calendar.list always showed icons in white. Now configured icon0 class is used.
 
 ### Known Bugs
 - if item contains a stringified number (e.g. with leading zero). widget.set converts it back to numeric format - so basic.print can not print it as text